Use a Money to represent and manipulate prices

+++ priv/repo/migrations/20170730090509_create_products.exs
@@ -5,7 +5,7 @@ defmodule Acme.Repo.Migrations.CreateProducts do
     create table(:products, primary_key: false) do
       add :id, :binary_id, primary_key: true
       add :name, :string
-      add :price, :float
+      add :price, :integer
 
       timestamps()
     end
